fixing comments.
